+++ src/runtime/composables/useGloriousFetch.ts
@@ -1,57 +1,63 @@
-import { useCookie, useFetch, useRuntimeConfig } from "nuxt/app";
-import { GloriousStore } from "../stores/GloriousStore";
-import defu from "defu";
+import { useCookie, useFetch, useRuntimeConfig } from 'nuxt/app'
+import { GloriousStore } from '../stores/GloriousStore'
+import defu from 'defu'
 interface gloriousFetchOptions {
-  gKey?: String;
-  params?: Object;
-  server?: Boolean;
-  is$?: Boolean;
-  lazy?: Boolean;
-  headers?: Object;
-  body?: Object;
-  bodyType?: "formData" | "formDataCustom" | "normal";
-  method?: "POST" | "GET" | "PATCH" | "PUT" | "DELETE" | "HEAD";
-  credentials?: "same-origin" | "include";
-  watch?: Array<Object>;
+  gKey?: String
+  params?: Object
+  server?: Boolean
+  is$?: Boolean
+  lazy?: Boolean
+  headers?: Object
+  body?: Object
+  keepResponse?: Boolean
+  bodyType?: 'formData' | 'formDataCustom' | 'normal'
+  method?: 'POST' | 'GET' | 'PATCH' | 'PUT' | 'DELETE' | 'HEAD'
+  credentials?: 'same-origin' | 'include'
+  watch?: Array<Object>
 }
 const defaultOptions: gloriousFetchOptions = {
   server: false,
-  method: "GET",
+  method: 'GET',
   lazy: true,
   is$: true,
+  keepResponse: false,
   params: {},
   headers: {
-    Accept: "application/json",
+    Accept: 'application/json',
   },
-  bodyType: "normal",
-  credentials: "same-origin",
-};
+  bodyType: 'normal',
+  credentials: 'same-origin',
+}
 
-export default function (url: string, options: gloriousFetchOptions = defaultOptions) {
-  const moduleConfig: any = useRuntimeConfig();
-  options = defu(moduleConfig.public.glorious.fetch, options, defaultOptions);
+export default function (
+  url: string,
+  options: gloriousFetchOptions = defaultOptions
+) {
+  const moduleConfig: any = useRuntimeConfig()
+  options = defu(moduleConfig.public.glorious.fetch, options, defaultOptions)
 
-  const gs: any = GloriousStore();
-  const gKey: String = computeGKey(options.gKey, url);
+  const gs: any = GloriousStore()
+  const gKey: String = computeGKey(options.gKey, url)
 
-  options.params = computeParams(<Object>options.params);
-  options.headers = { ...options.headers, ...computeAuth() };
+  options.params = computeParams(<Object>options.params)
+  options.headers = { ...options.headers, ...computeAuth() }
 
-  if (options.bodyType === "formData") options.body = computeFormData(options);
+  if (options.bodyType === 'formData') options.body = computeFormData(options)
 
   const opt: gloriousFetchOptions | any = {
     ...options,
     onRequest() {
       try {
-        gs.loading[<string>gKey] = true;
+        gs.loading[<string>gKey] = true
       } catch (e) {
         /* empty */
       }
     },
-    onResponse(/*{ response: res }:any*/) {
+    onResponse({ response }: any) {
+      gs.response[<string>gKey] = response._data
       try {
-        gs.loading[<string>gKey] = false;
-        gs.forms[<string>gKey].errors = [];
+        gs.loading[<string>gKey] = false
+        gs.forms[<string>gKey].errors = []
         // if (
         //   res.status >= 200 &&
         //   res.status <= 299 &&
@@ -63,85 +69,89 @@ export default function (url: string, options: gloriousFetchOptions = defaultOpt
         /* empty */
       }
     },
-    async onResponseError({ response: res }:any) {
-      const fetch = import.meta.glob("/glorious/fetch.ts");
-      if (typeof fetch["/glorious/fetch.ts"] !== "undefined")
-        fetch["/glorious/fetch.ts"]().then((data: any) => {
-          data.fetchHandler.onResponseError(res, gKey);
-        });
+    async onResponseError({ response: res }: any) {
+      const fetch = import.meta.glob('/glorious/fetch.ts')
+      if (typeof fetch['/glorious/fetch.ts'] !== 'undefined')
+        fetch['/glorious/fetch.ts']().then((data: any) => {
+          data.fetchHandler.onResponseError(res, gKey)
+        })
       else {
         if (res.status === 422) {
           try {
-            gs.forms[<string>gKey].errors = res._data.errors;
+            gs.forms[<string>gKey].errors = res._data.errors
           } catch (e) {
             /* empty */
           }
         }
-        if (res.status === 401 && process.client) gs.authLogout();
+        if (res.status === 401 && process.client) gs.authLogout()
       }
     },
-  };
+  }
 
-  if (opt.method === "GET" && typeof opt.body !== "undefined") {
-    opt.method = "POST";
-    return $fetch(url, opt);
+  if (opt.keepResponse && !gs.keepResponse.includes(gKey))
+    gs.keepResponse.push(gKey)
+
+  if (opt.method === 'GET' && typeof opt.body !== 'undefined') {
+    opt.method = 'POST'
+    return $fetch(url, opt)
   } else if (
-    opt.method === "GET" &&
-    typeof opt.body === "undefined" &&
+    opt.method === 'GET' &&
+    typeof opt.body === 'undefined' &&
     !opt.is$
   )
-    return useFetch(url, opt);
-  else return $fetch(url, opt);
+    return useFetch(url, opt)
+  else return $fetch(url, opt)
 }
 
 function computeParams(params: Object): Object {
-  const computeParams: any = {};
+  const computeParams: any = {}
   Object.entries(params).map((item: any) => {
-    if (item[1] !== null && item[1] !== "") computeParams[item[0]] = item[1];
-  });
+    if (item[1] !== null && item[1] !== '') computeParams[item[0]] = item[1]
+  })
 
-  return computeParams;
+  return computeParams
 }
 
 function computeGKey(gKey: String | undefined, url: string) {
-  return typeof gKey !== "undefined"
+  return typeof gKey !== 'undefined'
     ? gKey
-    : url.split("/")[url.split("/").length - 1];
+    : url.split('/')[url.split('/').length - 1]
 }
 
 function computeAuth(): Object {
-  const moduleConfig: any = useRuntimeConfig();
+  const moduleConfig: any = useRuntimeConfig()
 
-  type headerType = { Authorization?: String };
-  const header: headerType = {};
-  const token = useCookie(moduleConfig.public.glorious.auth.cookie.name);
+  type headerType = { Authorization?: String }
+  const header: headerType = {}
+  const token = useCookie(moduleConfig.public.glorious.auth.cookie.name)
 
-  if (typeof token.value !== "undefined")
-    header.Authorization = "Bearer " + token.value;
+  if (typeof token.value !== 'undefined')
+    header.Authorization = 'Bearer ' + token.value
 
-  return header;
+  return header
 }
 
 function computeFormData(options: gloriousFetchOptions) {
-  const form: any = new FormData();
+  const form: any = new FormData()
 
   Object.entries(<Object>options.body).forEach((item: any) => {
-    console.log(item[1]);
+    console.log(item[1])
 
-    if (item[1] === null) return;
+    if (item[1] === null) return
 
     if (
-      typeof item[1] === "object" &&
-      typeof item[1].lastModifiedDate === "undefined" &&
-      (typeof item[1].type === 'undefined' && typeof item[1].size === 'undefined')
+      typeof item[1] === 'object' &&
+      typeof item[1].lastModifiedDate === 'undefined' &&
+      typeof item[1].type === 'undefined' &&
+      typeof item[1].size === 'undefined'
     )
       Object.entries(item[1]).forEach((nestedItem, index) => {
-        form.append(`${item[0]}[${index}]`, nestedItem[1]);
-      });
+        form.append(`${item[0]}[${index}]`, nestedItem[1])
+      })
     else {
-      form.append(`${item[0]}`, item[1]);
+      form.append(`${item[0]}`, item[1])
     }
-  });
+  })
 
-  return form;
+  return form
 }